# packages/shared-types/api/openapi.json
{"openapi":"3.1.0","info":{"title":"DS Bridge AI Server","description":"\n## DS Bridge AI Server\n\n디자인 시스템 컴포넌트를 기반으로 React UI 코드를 생성하는 AI 서버입니다.\n\n### 인증\n\n모든 API는 `X-API-Key` 헤더 인증이 필요합니다. (`/health` 제외)\n```\nX-API-Key: sk-your-secret-key\n```\n\n### 주요 기능\n\n- **AI 채팅**: 자연어로 UI를 설명하면 React TSX 코드 생성\n- **멀티 프로바이더**: OpenAI, Anthropic, Gemini 지원\n- **스트리밍**: SSE 기반 실시간 응답\n- **컴포넌트 인식**: 디자인 시스템 스키마 기반 정확한 코드 생성\n- **동적 스키마**: Firebase Storage에서 컴포넌트 스키마 로드 지원\n\n### 응답 형식 (Streaming)\n\n```\ndata: {\"type\": \"chat\", \"text\": \"모던한 로그인 폼입니다.\"}\ndata: {\"type\": \"code\", \"path\": \"src/pages/Login.tsx\", \"content\": \"...\"}\ndata: {\"type\": \"done\"}\n```\n","version":"0.1.0","x-logo":{"url":"https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"}},"paths":{"/room/create":{"post":{"tags":["room"],"summary":"채팅방 생성","description":"새로운 채팅방을 생성합니다.\n\n## 요청 예시\n```json\n{\n  \"storybook_url\": \"https://storybook.example.com\",\n  \"user_id\": \"user-123\"\n}\n```\n\n## 응답\n- `id`: 생성된 채팅방 UUID\n- `storybook_url`: Storybook URL\n- `user_id`: 사용자 ID\n- `created_at`: 생성 시간","operationId":"create_room_room_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomRequest"}}},"required":true},"responses":{"201":{"description":"채팅방 생성 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}},"500":{"description":"서버 오류"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"X-API-Key":[]}]}},"/room/get/{room_id}":{"get":{"tags":["room"],"summary":"채팅방 조회","description":"채팅방 ID로 채팅방 정보를 조회합니다.","operationId":"get_room_room_get__room_id__get","security":[{"X-API-Key":[]}],"parameters":[{"name":"room_id","in":"path","required":true,"schema":{"type":"string","title":"Room Id"}}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}},"404":{"description":"채팅방을 찾을 수 없음"},"500":{"description":"서버 오류"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/send":{"post":{"tags":["chat"],"summary":"AI 채팅 (Non-streaming)","description":"자연어로 UI를 설명하면 React TSX 코드를 생성합니다.\n\n## 요청 예시\n```json\n{\n  \"message\": \"로그인 페이지 만들어줘\",\n  \"room_id\": \"550e8400-e29b-41d4-a716-446655440000\",\n  \"schema_key\": \"schemas/component-schema.json\"\n}\n```\n\n## 응답 구조\n- `message`: AI 원본 응답\n- `parsed.conversation`: 한국어 설명\n- `parsed.files`: 생성된 코드 파일 목록\n- `usage`: 토큰 사용량\n\n## room_id\n채팅방 ID (필수). 메시지가 해당 채팅방에 저장됩니다.\n\n## schema_key\nFirebase Storage에서 컴포넌트 스키마를 로드합니다. 생략 시 로컬 스키마 사용.","operationId":"chat_chat_send_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"400":{"description":"잘못된 요청 (stream=true인 경우)"},"404":{"description":"채팅방을 찾을 수 없음"},"500":{"description":"AI API 호출 실패"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"X-API-Key":[]}]}},"/chat/stream":{"post":{"tags":["chat"],"summary":"AI 채팅 (Streaming)","description":"SSE(Server-Sent Events)를 통해 실시간 스트리밍 응답을 받습니다.\n\n## 요청 예시\n```json\n{\n  \"message\": \"로그인 페이지 만들어줘\",\n  \"room_id\": \"550e8400-e29b-41d4-a716-446655440000\",\n  \"schema_key\": \"schemas/component-schema.json\"\n}\n```\n\n## 이벤트 타입\n\n| 타입 | 설명 | 필드 |\n|------|------|------|\n| `chat` | 대화 텍스트 (실시간) | `text` |\n| `code` | 코드 파일 (완성 후) | `path`, `content` |\n| `done` | 스트리밍 완료 | - |\n| `error` | 오류 발생 | `error` |\n\n## SSE 응답 예시\n```\ndata: {\"type\": \"chat\", \"text\": \"모던한 \"}\n\ndata: {\"type\": \"chat\", \"text\": \"로그인 페이지입니다.\"}\n\ndata: {\"type\": \"code\", \"path\": \"src/pages/Login.tsx\", \"content\": \"import...\"}\n\ndata: {\"type\": \"done\"}\n```\n\n## room_id\n채팅방 ID (필수). 스트리밍 완료 후 메시지가 해당 채팅방에 저장됩니다.","operationId":"chat_stream_chat_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"SSE 스트림","content":{"application/json":{"schema":{}},"text/event-stream":{"example":"data: {\"type\": \"chat\", \"text\": \"모던한 로그인\"}\n\ndata: {\"type\": \"done\"}\n\n"}}},"404":{"description":"채팅방을 찾을 수 없음"},"500":{"description":"AI API 호출 실패"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"X-API-Key":[]}]}},"/component/schema":{"get":{"tags":["component"],"summary":"컴포넌트 스키마 조회","description":"현재 로드된 디자인 시스템 컴포넌트 스키마를 반환합니다.","operationId":"get_components_component_schema_get","responses":{"200":{"description":"스키마 조회 성공","content":{"application/json":{"schema":{},"example":{"version":"1.0.0","generatedAt":"2026-01-09T10:00:00.000Z","components":{"Button":{"displayName":"Button","category":"UI","props":{"variant":{"type":["primary","secondary"],"required":false}}}}}}}},"404":{"description":"스키마 파일을 찾을 수 없음"}},"security":[{"X-API-Key":[]}]}},"/component/reload":{"post":{"tags":["component"],"summary":"컴포넌트 스키마 리로드","description":"component-schema.json 파일을 다시 로드하여 시스템 프롬프트를 갱신합니다.","operationId":"reload_components_component_reload_post","responses":{"200":{"description":"리로드 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReloadResponse"}}}},"500":{"description":"스키마 파일 로드 실패"}},"security":[{"X-API-Key":[]}]}},"/health":{"get":{"tags":["health"],"summary":"서버 상태 확인","description":"서버가 정상적으로 실행 중인지 확인합니다.","operationId":"health_check_health_get","responses":{"200":{"description":"서버 정상","content":{"application/json":{"schema":{},"example":{"status":"healthy"}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"message":{"type":"string","title":"Message","description":"사용자 메시지","example":"로그인 페이지 만들어줘"},"room_id":{"type":"string","title":"Room Id","description":"채팅방 ID","example":"550e8400-e29b-41d4-a716-446655440000"},"stream":{"type":"boolean","title":"Stream","description":"스트리밍 응답 여부 (True면 /stream 엔드포인트 사용 권장)","default":false},"schema_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Schema Key","description":"Firebase Storage 내 컴포넌트 스키마 경로 (예: schemas/v1/component-schema.json)","example":"schemas/v1/component-schema.json"}},"type":"object","required":["message","room_id"],"title":"ChatRequest","description":"채팅 요청","examples":[{"message":"로그인 페이지 만들어줘","room_id":"550e8400-e29b-41d4-a716-446655440000"},{"message":"대시보드 만들어줘","room_id":"550e8400-e29b-41d4-a716-446655440000","schema_key":"schemas/v1/component-schema.json"}]},"ChatResponse":{"properties":{"message":{"$ref":"#/components/schemas/Message","description":"AI 응답 메시지"},"parsed":{"anyOf":[{"$ref":"#/components/schemas/ParsedResponse"},{"type":"null"}],"description":"파싱된 응답 (conversation과 files 분리)"},"usage":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Usage","description":"토큰 사용량 정보","example":{"completion_tokens":800,"prompt_tokens":1500,"total_tokens":2300}}},"type":"object","required":["message"],"title":"ChatResponse","description":"채팅 응답 (Non-streaming)","examples":[{"message":{"content":"모던한 로그인 페이지입니다.\n\n<file path=\"src/pages/Login.tsx\">import { Button } from '@/components';\n\nconst Login = () => <Button>로그인</Button>;\n\nexport default Login;</file>","role":"assistant"},"parsed":{"conversation":"모던한 로그인 페이지입니다.","files":[{"content":"import { Button } from '@/components';\n\nconst Login = () => <Button>로그인</Button>;\n\nexport default Login;","path":"src/pages/Login.tsx"}],"raw":"..."},"usage":{"completion_tokens":800,"prompt_tokens":1500,"total_tokens":2300}}]},"CreateRoomRequest":{"properties":{"storybook_url":{"type":"string","title":"Storybook Url","description":"Storybook URL","example":"https://storybook.example.com"},"user_id":{"type":"string","title":"User Id","description":"사용자 ID","example":"user-123"}},"type":"object","required":["storybook_url","user_id"],"title":"CreateRoomRequest","description":"채팅방 생성 요청","examples":[{"storybook_url":"https://storybook.example.com","user_id":"user-123"}]},"FileContent":{"properties":{"path":{"type":"string","title":"Path","description":"파일 경로 (예: src/pages/Login.tsx)","example":"src/pages/Login.tsx"},"content":{"type":"string","title":"Content","description":"파일 내용 (React TSX 코드)","example":"import { Button } from '@/components';\n\nconst Login = () => {\n  return <Button>로그인</Button>;\n};\n\nexport default Login;"}},"type":"object","required":["path","content"],"title":"FileContent","description":"생성된 파일 정보","examples":[{"content":"import { Button, Field } from '@/components';\n\nconst Login = () => {\n  return (\n    <div>\n      <Field label=\"이메일\" />\n      <Button>로그인</Button>\n    </div>\n  );\n};\n\nexport default Login;","path":"src/pages/Login.tsx"}]},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Message":{"properties":{"role":{"type":"string","enum":["user","assistant","system"],"title":"Role","description":"메시지 역할","example":"user"},"content":{"type":"string","title":"Content","description":"메시지 내용","example":"로그인 페이지 만들어줘"}},"type":"object","required":["role","content"],"title":"Message","description":"채팅 메시지","examples":[{"content":"로그인 페이지 만들어줘","role":"user"},{"content":"모던한 로그인 페이지입니다.\n\n<file path=\"src/pages/Login.tsx\">...</file>","role":"assistant"}]},"ParsedResponse":{"properties":{"conversation":{"type":"string","title":"Conversation","description":"AI의 한국어 설명 (코드 외 텍스트)","example":"모던하고 깔끔한 로그인 페이지입니다. 그라데이션 배경과 카드 레이아웃을 사용했습니다."},"files":{"items":{"$ref":"#/components/schemas/FileContent"},"type":"array","title":"Files","description":"생성된 코드 파일 목록"},"raw":{"type":"string","title":"Raw","description":"AI 원본 응답 (파싱 전)"}},"type":"object","required":["conversation","raw"],"title":"ParsedResponse","description":"파싱된 AI 응답 - 대화와 코드 파일 분리","examples":[{"conversation":"모던한 로그인 페이지입니다.","files":[{"content":"import { Button } from '@/components';\n...","path":"src/pages/Login.tsx"}],"raw":"모던한 로그인 페이지입니다.\n\n<file path=\"src/pages/Login.tsx\">...</file>"}]},"ReloadResponse":{"properties":{"message":{"type":"string","title":"Message","description":"결과 메시지","example":"Schema reloaded successfully"},"componentCount":{"type":"integer","title":"Componentcount","description":"리로드된 컴포넌트 수","example":25}},"type":"object","required":["message","componentCount"],"title":"ReloadResponse","description":"스키마 리로드 응답"},"RoomResponse":{"properties":{"id":{"type":"string","title":"Id","description":"채팅방 ID (UUID)","example":"550e8400-e29b-41d4-a716-446655440000"},"storybook_url":{"type":"string","title":"Storybook Url","description":"Storybook URL","example":"https://storybook.example.com"},"user_id":{"type":"string","title":"User Id","description":"사용자 ID","example":"user-123"},"created_at":{"type":"integer","title":"Created At","description":"생성 시간 (ms timestamp)","example":1736654400000}},"type":"object","required":["id","storybook_url","user_id","created_at"],"title":"RoomResponse","description":"채팅방 응답"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"X-API-Key":{"type":"apiKey","in":"header","name":"X-API-Key","description":"API 키 인증. Chat API 엔드포인트에 필요합니다."}}},"tags":[{"name":"health","description":"서버 상태 확인 엔드포인트"},{"name":"room","description":"채팅방 관리 API"},{"name":"chat","description":"AI 채팅 API - 디자인 시스템 컴포넌트 기반 React UI 코드 생성"},{"name":"component","description":"컴포넌트 스키마 관리 API"}]}